refactor(mobile): merge search into Discover tab
Move search UI directly into the Discover tab (watch.tsx) instead of
a separate pushed screen. The Discover tab now shows the search bar
and results inline — search IS the discover experience.

- Remove standalone app/search.tsx route and its Stack.Screen
- Remove search icon from Discover tab header (no longer needed)
- Keep Discover header with title for visual consistency
