+++ apps/mobile/app/_layout.tsx
@@ -243,31 +243,6 @@ export default function RootLayout() {
                     ),
                   }}
                 />
-                <Stack.Screen
-                  name="search"
-                  options={{
-                    headerShown: true,
-                    headerTintColor: ACCENT,
-                    headerTitle: "Search",
-                    headerStyle: { backgroundColor: BG_COLOR },
-                    headerShadowVisible: false,
-                    headerTitleAlign: "center",
-                    headerLeft: () => (
-                      <Pressable
-                        onPress={() => router.back()}
-                        accessibilityRole="button"
-                        accessibilityLabel="Go back"
-                        hitSlop={12}
-                      >
-                        <Ionicons
-                          name="chevron-back"
-                          size={28}
-                          color={ACCENT}
-                        />
-                      </Pressable>
-                    ),
-                  }}
-                />
               </Stack>
             </ExperienceShell>
           </ExperienceSelectionProvider>
